增加了冻结功能，当被冻结的时候，不会删除运行过的结果文件。

+++ main.py
@@ -1,9 +1,11 @@
+# coding=UTF-8
 from generate_ast import generate_ast
 from read_ast import read_ast
 from remove_annotation import remove_annotation
 import os
 import shutil
 import datetime
+import config
 from train import train
 
 if __name__ == '__main__':
@@ -14,9 +16,10 @@ if __name__ == '__main__':
         os.mkdir("/home/xjj/AST-GNN/data/sol_source/")
     if not os.path.exists("/home/xjj/AST-GNN/data/AST_json/"):
         os.mkdir("/home/xjj/AST-GNN/data/AST_json/")
-    # 在运行之前先删除之前轮次运行出来的结果
-    for project_name in os.listdir("/home/xjj/AST-GNN/data/AST_json/"):
-        shutil.rmtree("/home/xjj/AST-GNN/data/AST_json/" + project_name)
+    # 在运行之前先删除之前轮次运行出来的结果,这里当初只是为了方便dev的时候写的。
+    if config.frozen == "delete":
+        for project_name in os.listdir("/home/xjj/AST-GNN/data/AST_json/"):
+            shutil.rmtree("/home/xjj/AST-GNN/data/AST_json/" + project_name)
     # 删除源代码中的注释
     remove_annotation()
     # 先利用solc的编译器，直接将源代码转化为抽象语法树的AST文件
